requirements.txt
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bson==0.5.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
docx==0.2.4
executing==2.0.1
Flask==3.0.3
fonttools==4.53.1
functions-framework==3.8.1
h11==0.14.0
idna==3.10
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.3
pillow==10.4.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
pymongo==4.6.2
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0
python-docx==1.1.2
pytz==2024.1
pywin32==306
pyzmq==25.1.2
requests==2.32.3
scipy==1.14.1
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
